Refactor : 클래스 위치 수정
- 소켓 관련 클래스들의 위치를 수정합니다.
